Create menu prompt
We've created a dedicated prompt for selecting a certain menu id.
