+++ lib/cli/prompts/index.js
@@ -12,26 +12,9 @@ export { default as number } from '@inquirer/number';
 export { default as select } from '@inquirer/select';
 export { default as checkbox, Separator } from '@inquirer/checkbox';
 export { default as nestedList } from './nested-list.js';
+export * from './hex-prompt.js';
+export * from './menu-prompt.js';
 export { input, fileSelector, confirm };
-
-// # hex(opts)
-// Helper function for requesting a single hexadecimal number as input, with 
-// validation.
-export function hex(opts) {
-	return input({
-		validate(input) {
-			if (!input.startsWith('0x')) {
-				return 'Please input a valid hexadecimal number';
-			}
-			let nr = +input;
-			if (Number.isNaN(nr)) {
-				return 'Please input a valid hexadecimal number';
-			}
-			return true;
-		},
-		...opts,
-	});
-}
 
 // # uint32(opts)
 // Helper function for requesting a simple uint32 numerical value.
